+++ webpack.config.js
@@ -106,7 +106,38 @@ module.exports = (env, argv) => {
                     ]
                 },
                 {
+                    test: /\.svg$/i,
+                    include: [
+                        path.resolve(__dirname, "src/icons/flags/1x1")
+                    ],
+                    use: [
+                        {
+                            loader: "file-loader",
+                            options: {
+                                outputPath: "img/flags/1x1/",
+                                name: "[name].[ext]?[hash]"
+                            }
+                        }
+                    ]
+                },
+                {
+                    test: /\.svg$/i,
+                    include: [
+                        path.resolve(__dirname, "src/icons/flags/4x3")
+                    ],
+                    use: [
+                        {
+                            loader: "file-loader",
+                            options: {
+                                outputPath: "img/flags/4x3/",
+                                name: "[name].[ext]?[hash]"
+                            }
+                        }
+                    ]
+                },
+                {
                     test: /\.(png|jpe?g|gif|svg)$/i,
+                    exclude: /flags/,
                     use: [
                         {
                             loader: "file-loader",